Worked on the flow from my collection to review page

+++ Android/app/src/main/java/com/marmarovas/animediary/network/animes/AnimeData.kt
@@ -3,6 +3,6 @@ package com.marmarovas.animediary.network.animes
 data class AnimeData (
     val anime : Anime,
     val notes : String?,
-    val rating : Int?,
+    val rating : Float?,
     val tags : List<String>?
 )
